# Conflicts:
#	maxtempmapreduce-bigdata/src/main/java/WeatherMapper.java

+++ maxtempmapreduce-bigdata/src/main/java/WeatherReducer.java
@@ -1,4 +1,5 @@
 import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.MapReduceBase;
 import org.apache.hadoop.mapred.OutputCollector;
@@ -8,19 +9,19 @@ import org.apache.hadoop.mapred.Reporter;
 import java.io.IOException;
 import java.util.Iterator;
 
-public class WeatherReducer extends MapReduceBase implements Reducer<Text, DoubleWritable, Text, DoubleWritable> {
+public class WeatherReducer extends MapReduceBase implements Reducer<Text, IntWritable, Text, IntWritable> {
 
-    public void reduce(Text text, Iterator<DoubleWritable> iterator, OutputCollector<Text, DoubleWritable> outputCollector, Reporter reporter) throws IOException {
-        double maxTemperature= 0;
+    public void reduce(Text text, Iterator<IntWritable> iterator, OutputCollector<Text, IntWritable> outputCollector, Reporter reporter) throws IOException {
+        int maxTemperature= 0;
 
         while(iterator.hasNext()) {
-            double curr = iterator.next().get();
+            int curr = iterator.next().get();
             if(maxTemperature < curr) {
                 maxTemperature = curr;
             }
         }
         //the temperature in the data-set is represented in celsius times 10
         maxTemperature = maxTemperature /10;
-        outputCollector.collect(text, new DoubleWritable(maxTemperature));
+        outputCollector.collect(text, new IntWritable(maxTemperature));
     }
 }
